+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 Upcoming
 --------
 
+- Fixed parsing of compact ``YYYYMMDD`` date strings (no separators) being
+  misread as a year value.
+
 v0.17.0
 -------